# docs/swagger.yaml
components:
  schemas:
...
    api.commentResponse:
      properties:
        author_avatar:
          type: string
        author_username:
          type: string
        content:
          type: string
        created_at:
          type: string
        id:
          type: string
        like_count:
          type: integer
        parent_id:
          type: string
        post_id:
          type: string
        user_id:
          type: string
      type: object
    api.createCommentRequest:
      properties:
        content:
          maxLength: 1000
          type: string
        parent_id:
          type: string
        post_id:
          type: string
      required:
      - content
      - post_id
      type: object
    api.createCommentResponse:
      properties:
        id:
          type: string
      type: object
...
    api.followResponse:
      properties:
        ok:
          type: boolean
      type: object
    api.followUserResponse:
      properties:
        avatar_url:
          type: string
        bio:
          type: string
        id:
          type: string
        username:
          type: string
      type: object
...
    render.Response-api_createCommentResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.createCommentResponse'
        msg:
          type: string
      type: object
...
    render.Response-api_followResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.followResponse'
        msg:
          type: string
      type: object
...
    render.Response-array_api_commentResponse:
      properties:
        code:
          type: integer
        data:
          items:
            $ref: '#/components/schemas/api.commentResponse'
          type: array
          uniqueItems: false
        msg:
          type: string
      type: object
    render.Response-array_api_followUserResponse:
      properties:
        code:
          type: integer
        data:
          items:
            $ref: '#/components/schemas/api.followUserResponse'
          type: array
          uniqueItems: false
        msg:
          type: string
      type: object
...
    render.errorResponse:
      properties:
        code:
          type: integer
        msg:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      scheme: bearer
      type: http
...
info:
  description: 小蓝书后端接口文档
  title: 小蓝书 API
  version: "1.0"
openapi: 3.1.0
paths:
  /comments:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.createCommentRequest'
              description: 评论内容
              summary: body
        description: 评论内容
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_createCommentResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 创建评论
      tags:
      - comments
...
  /posts/{id}/comments:
    get:
      parameters:
      - description: 帖子 ID
        in: query
        name: post_id
        required: true
        schema:
          type: string
      - description: 页码
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: 每页数量
        in: query
        name: page_size
        schema:
          default: 20
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-array_api_commentResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      summary: 获取帖子评论列表
      tags:
      - comments
...
  /users/{id}/follow:
    delete:
      parameters:
      - description: 目标用户 ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_followResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 取消关注
      tags:
      - follows
    post:
      parameters:
      - description: 目标用户 ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_followResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 关注用户
      tags:
      - follows
  /users/{id}/followers:
    get:
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: 页码
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: 每页数量
        in: query
        name: page_size
        schema:
          default: 20
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-array_api_followUserResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      summary: 获取粉丝列表
      tags:
      - follows
  /users/{id}/following:
    get:
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: 页码
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: 每页数量
        in: query
        name: page_size
        schema:
          default: 20
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-array_api_followUserResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      summary: 获取关注列表
      tags:
      - follows
...
servers:
- description: 本地开发环境
  url: http://localhost:8080/api/v1